[WTEL-6157] moved interface to types file
