+++ src/ui/modules/work-section/modules/call/components/call-transfer/call-transfer-container.vue
@@ -49,10 +49,11 @@ import useInfiniteScroll from '../../../../../../../app/composables/useInfiniteS
 import LookupItemContainer from '../../../_shared/components/lookup-item-container/lookup-item-container.vue';
 import EmptySearch from '../../../_shared/components/workspace-empty-search/components/empty-search.vue';
 import TransferLookupItem from '../../../_shared/components/lookup-item/transfer-lookup-item.vue';
+import { transferParams } from './types/transfer-tabs'
 
 interface CallTransferContainerProps {
   size?: string;
-  getData: (params: any) => Promise<any>;
+  getData: (params: transferParams) => Promise<any>;
   showStatus?: boolean;
   showTeamName?: boolean;
   showUserNameAvatar?: boolean;
