+++ src/ui/modules/work-section/modules/call/components/call-transfer/types/transfer-tabs.ts
@@ -1,0 +1,10 @@
+interface TransferParams {
+  page?: number;
+  size?: number;
+  q?: string;
+  sort?: string;
+  fields?: string[];
+  notId?: string[];
+  staus?: string;
+  [key: string]: any;
+}
